feat: add lib for request verification and catch for bad request
