+++ app/routes/login.ts
@@ -1,6 +1,6 @@
 import { ActionFunction } from "@remix-run/node";
 import { z } from "zod";
-import { verify } from "~/helpers/jwt-helper.server";
+import { verifyRequest } from "~/lib/verify-request.server";
 import { findUserByLogin } from "~/models/user.server";
 
 type ActionData = {
@@ -10,46 +10,45 @@ type ActionData = {
 export const action: ActionFunction = async ({
   request,
 }): Promise<ActionData> => {
-  const authorizationHeader = request.headers.get("Authorization");
-  if (authorizationHeader) {
-    const [_, jwt] = authorizationHeader.split(" ");
-    verify(jwt);
+  verifyRequest(request);
 
-    const body = await request.json();
-
-    const loginSchema = z.object({
-      email: z.string().email(),
-      // the password is hashed, so it must be at least 64 chars
-      password: z.string().min(64),
+  let body: any;
+  try {
+    body = await request.json();
+  } catch (e) {
+    throw new Response(null, {
+      status: 400,
+      statusText: "Bad Request",
     });
+  }
 
-    const parsedData = loginSchema.safeParse(body);
+  const loginSchema = z.object({
+    email: z.string().email(),
+    // the password is hashed, so it must be at least 64 chars
+    password: z.string().min(64),
+  });
 
-    if (parsedData.success) {
-      const data = parsedData.data;
+  const parsedData = loginSchema.safeParse(body);
 
-      const user = await findUserByLogin(data.email, data.password);
+  if (parsedData.success) {
+    const data = parsedData.data;
 
-      if (user) {
-        return {
-          id: user.id,
-        };
-      } else {
-        throw new Response(null, {
-          status: 404,
-          statusText: "User not found",
-        });
-      }
+    const user = await findUserByLogin(data.email, data.password);
+
+    if (user) {
+      return {
+        id: user.id,
+      };
     } else {
       throw new Response(null, {
-        status: 400,
-        statusText: "Bad Request",
+        status: 404,
+        statusText: "User not found",
       });
     }
   } else {
     throw new Response(null, {
-      status: 401,
-      statusText: "Unauthorized",
+      status: 400,
+      statusText: "Bad Request",
     });
   }
 };
